requirements.txt
openai
tiktoken
numpy
sqlite_utils
pycryptodome
matplotlib
plotly